# apps/ui/swagger.yaml
openapi: 3.0.0
paths:
  /auth/register:
    post:
      operationId: register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDTO'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '409':
          content:
            application/json:
              examples:
                UsernameConflictException:
                  description: User with such username already exists
                  value:
                    statusCode: 409
                    message: User with such username already exists
                    error: Conflict
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 409
                  message:
                    type: string
                    example: User with such username already exists
                  error:
                    type: string
                    example: Conflict
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Auth
  /auth/login:
    post:
      operationId: login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                InvalidPasswordException:
                  description: Invalid password
                  value:
                    statusCode: 401
                    message: Invalid password
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Invalid password
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
        '404':
          content:
            application/json:
              examples:
                UserNotFoundException:
                  description: User not found
                  value:
                    statusCode: 404
                    message: User not found
                    error: Not Found
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: User not found
                  error:
                    type: string
                    example: Not Found
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Auth
  /auth/logout:
    delete:
      operationId: logout
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: boolean
      tags:
        - Auth
  /auth/profile:
    get:
      operationId: profile
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Auth
    delete:
      operationId: deleteAccount
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: boolean
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Auth
  /auth/username/{username}:
    get:
      operationId: checkUsername
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsernameCheckDTO'
      tags:
        - Auth
  /links:
    post:
      operationId: createLink
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkDTO'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
        '409':
          content:
            application/json:
              examples:
                BackHalfIsNotUniqueException:
                  description: Backhalf is not unique
                  value:
                    statusCode: 409
                    message: Backhalf is not unique
                    error: Conflict
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 409
                  message:
                    type: string
                    example: Backhalf is not unique
                  error:
                    type: string
                    example: Conflict
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Links
    get:
      operationId: findAllLinks
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            type: number
        - name: perPage
          required: false
          in: query
          schema:
            default: 20
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Link'
                      meta:
                        type: object
                        properties:
                          itemCount:
                            type: number
                          totalItems:
                            type: number
                          itemsPerPage:
                            type: number
                          totalPages:
                            type: number
                          currentPage:
                            type: number
                        required:
                          - itemCount
                          - totalItems
                          - itemsPerPage
                          - totalPages
                          - currentPage
                      links:
                        type: object
                        properties:
                          first:
                            type: string
                          previous:
                            type: string
                          next:
                            type: string
                          last:
                            type: string
                        required:
                          - first
                          - previous
                          - next
                          - last
                    required:
                      - items
                      - meta
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Links
  /links/{linkId}:
    post:
      operationId: updateLink
      parameters:
        - name: linkId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLinkDTO'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
        '404':
          content:
            application/json:
              examples:
                LinkNotFoundException:
                  description: Link not found
                  value:
                    statusCode: 404
                    message: Link not found
                    error: Not Found
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Link not found
                  error:
                    type: string
                    example: Not Found
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Links
    delete:
      operationId: deleteLink
      parameters:
        - name: linkId
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: boolean
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
        '404':
          content:
            application/json:
              examples:
                LinkNotFoundException:
                  description: Link not found
                  value:
                    statusCode: 404
                    message: Link not found
                    error: Not Found
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Link not found
                  error:
                    type: string
                    example: Not Found
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Links
    get:
      operationId: findLinkById
      parameters:
        - name: linkId
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
        '404':
          content:
            application/json:
              examples:
                LinkNotFoundException:
                  description: Link not found
                  value:
                    statusCode: 404
                    message: Link not found
                    error: Not Found
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Link not found
                  error:
                    type: string
                    example: Not Found
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Links
  /links/{linkId}/statistic:
    get:
      operationId: linkStatistic
      parameters:
        - name: linkId
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkStatisticDTO'
        '400':
          content:
            application/json:
              examples:
                BadRequestException:
                  description: Bad Request
                  value:
                    statusCode: 400
                    message: Bad Request
                    error: Bad Request
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Bad Request
                  error:
                    type: string
                    example: Bad Request
                required:
                  - statusCode
                  - message
          description: ''
        '401':
          content:
            application/json:
              examples:
                UnauthorizedException:
                  description: Unauthorized
                  value:
                    statusCode: 401
                    message: Unauthorized
                    error: Unauthorized
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 401
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: Unauthorized
                required:
                  - statusCode
                  - message
          description: ''
        '404':
          content:
            application/json:
              examples:
                LinkNotFoundException:
                  description: Link not found
                  value:
                    statusCode: 404
                    message: Link not found
                    error: Not Found
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Link not found
                  error:
                    type: string
                    example: Not Found
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Links
  /{backHalf}:
    get:
      operationId: click
      parameters:
        - name: backHalf
          required: true
          in: path
          schema:
            type: string
      responses:
        '301':
          description: Redirect to links destination.
        '404':
          content:
            application/json:
              examples:
                LinkNotFoundException:
                  description: Link not found
                  value:
                    statusCode: 404
                    message: Link not found
                    error: Not Found
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Link not found
                  error:
                    type: string
                    example: Not Found
                required:
                  - statusCode
                  - message
          description: ''
      tags:
        - Click
  /demo:
    post:
      operationId: demo
      summary: ''
      description: Creates a demo account with demo data
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - Demo
info:
  title: Link Shortener
  description: ''
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: sid
  schemas:
    RegisterDTO:
      type: object
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 32
        password:
          type: string
          minLength: 8
          maxLength: 32
      required:
        - username
        - password
    User:
      type: object
      properties:
        username:
          type: string
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - username
        - id
        - createdAt
        - updatedAt
    LoginDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    UsernameCheckDTO:
      type: object
      properties:
        isTaken:
          type: boolean
      required:
        - isTaken
    CreateLinkDTO:
      type: object
      properties:
        destination:
          type: string
          description: Url with protocol, host and tld.
          example: https://google.com
        name:
          type: string
          minLength: 1
          maxLength: 32
      required:
        - destination
        - name
    Link:
      type: object
      properties:
        backhalf:
          type: string
        name:
          type: string
        destination:
          type: string
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - backhalf
        - name
        - destination
        - id
        - createdAt
        - updatedAt
    UpdateLinkDTO:
      type: object
      properties:
        destination:
          type: string
          description: Url with protocol, host and tld.
          example: https://google.com
        name:
          type: string
          minLength: 1
          maxLength: 32
    LinkStatisticDTO:
      type: object
      properties:
        platforms:
          type: object
          additionalProperties:
            type: number
        browsers:
          type: object
          additionalProperties:
            type: number
        systems:
          type: object
          additionalProperties:
            type: number
        devices:
          type: object
          additionalProperties:
            type: number
        clicks:
          type: number
      required:
        - platforms
        - browsers
        - systems
        - devices
        - clicks